Define the entry model.
- The new class (Entry) inherits from Django's base Model class.
- The first attribute, sport, is a ForeignKey instance (a database term);
a code that connects each entry to a specific sport, as each sport record
is assigned a key, or ID, when it's created.
- The on_delete=models.CASCADE argument tells Django that when a sport is deleted,
all the entries associated with that sport should be deleted as well.
- The text attribute is an instance of TextField; no need for the size limit,
because we don't want to limit the size of individual entries.
- Meta class inside the Entry class holds extra information for managing our model.
It allows us to set a special attribute telling Django to use Entries (instead of Entrys) when referring to more
than one entry.
- The __str__() method tells Django the information to show when it refers to individual entries. Here we tell Django
to show just the first 50 characters of text.